gameboard & Boocard
gameboard is empty atm and boocard had psudo code because we need the gameboard 1st

+++ BoardGame/Assets/dice.cs
@@ -7,7 +7,7 @@ public class dice : MonoBehaviour
     // Start is called before the first frame update
     // creating a random number... will move to dice class
 
-    [SerializeField] float movementValue;
+    [SerializeField] int movementValue;
      public bool positive;
      [SerializeField]
